move code out of base when possible
